added apply_grid_ufunc method to Grid
